requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==3.0.8
django-haystack==2.8.1
django-redis==4.12.1
django-tinymce==2.7.0
dnspython==1.16.0
eventlet==0.26.1
fdfs-client-py==1.2.6
future==0.18.2
greenlet==0.4.16
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
jieba==0.42.1
kombu==4.6.11
monotonic==1.5
mutagen==1.45.0
Pillow==7.2.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
Whoosh==2.7.4
zipp==3.1.0
